% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_capped_ols.R
\name{binary_capped_ols}
\alias{binary_capped_ols}
\title{Capped OLS Regression}
\usage{
binary_capped_ols()
}
\value{
A linear regression model for binary outcomes,
with class 'learner'. Predictions are capped to fall inside [0,1].
}
\description{
Capped OLS Regression
}
