% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_convex_combination.R
\name{binary_convex_combo}
\alias{binary_convex_combo}
\title{Convex Combination for Regression}
\usage{
binary_convex_combo()
}
\value{
A model that fits a weighted average of the inputs, with weights
all nonnegative and summing to 1.
}
\description{
Convex Combination for Regression
}
\details{
Does not use weights.
}
