% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_elasticgbm.R
\name{binary_elasticgbm}
\alias{binary_elasticgbm}
\title{Elasticnet, then Boosted Trees}
\usage{
binary_elasticgbm(
  alpha = seq(0, 1, length.out = 5),
  learning_rate = 0.2,
  num_leaves = 7,
  max_depth = -1,
  min_data_in_leaf = 20,
  force_col_wise = FALSE,
  metric = "rmse",
  nrounds = 10000L,
  elastic_workers = 1L,
  lgbm_workers = 1L
)
}
\arguments{
\item{alpha}{sequence of alpha values to search over}

\item{learning_rate}{learning rate.}

\item{max_depth}{maximum depth of a tree.}

\item{elastic_workers}{integer number of cores to use for parallelized tuning of elasticnet}

\item{lgbm_workers}{integer number of cores to use for parallelized tuning of trees}
}
\value{
A lostic elasticnet + gradient boosting model for binary outcomes,
with class 'learner'.
}
\description{
Logistic elasticnet regression, with gradient boosted trees to pick the residual
}
