% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_elasticnet.R
\name{binary_elasticnet}
\alias{binary_elasticnet}
\title{Logistic elasticnet Regression}
\usage{
binary_elasticnet(alpha = seq(0, 1, length.out = 5))
}
\arguments{
\item{alpha}{sequence of alpha values to search over}
}
\value{
A logistic elasticnet regression model for binary outcomes,
with class 'learner'.
}
\description{
Construct a elasticnet regression model with glmnet.
}
