% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_lgbm.R
\name{binary_lgbm}
\alias{binary_lgbm}
\title{Gradient Boosted Tree Classifier}
\usage{
binary_lgbm(
  learning_rate = 0.05,
  max_depth = c(0:6),
  bagging_fraction = 1,
  feature_fraction = 1,
  num_threads = 1
)
}
\arguments{
\item{learning_rate}{learning rate.}

\item{max_depth}{maximum depth of a tree.}

\item{bagging_fraction}{bagging_fraction ratio of observations when constructing each
tree. Values less than 1 will speed up computation, and may reduce
overfitting.}

\item{feature_fraction}{bagging_fraction ratio of columns when constructing
each tree. Values less than 1 will speed up computation, and may reduce
overfitting.}

\item{num_threads}{number of threads to use, to speed up training and prediction}
}
\value{
A gradient boosted tree model for binary outcomes,
with class 'learner'.
}
\description{
Construct a gradient boosted tree classifier using lightgbm
All parameters can be passed as single values, or as vectors. Performs
grid search over all combinations of provided values, using early stopping
to select the best number of boosting iterations.
}
