% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_logistic.R
\name{binary_logistic}
\alias{binary_logistic}
\title{Logistic Regression}
\usage{
binary_logistic(engine = stats::glm.fit)
}
\value{
A logistic regression model for binary outcomes,
with class 'learner'.
}
\description{
Logistic Regression
}
