% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_random.R
\name{binary_random}
\alias{binary_random}
\title{Random Noise for Classification}
\usage{
binary_random(seed = 1)
}
\arguments{
\item{seed}{Random seed for replicability.}
}
\value{
A model that makes random predictions for a binary outcome
}
\description{
Random Noise for Classification
}
