% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_sampmean.R
\name{binary_sampmean}
\alias{binary_sampmean}
\title{Sample Mean for Classification}
\usage{
binary_sampmean()
}
\value{
A model that predicts the weighted sample mean
}
\description{
Sample Mean for Classification
}
