% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_stacking.R
\name{binary_stacking}
\alias{binary_stacking}
\title{Stacking for Classification}
\usage{
binary_stacking(learners = list(elasticnet = binary_elasticnet(), gbt =
  binary_gbt()))
}
\arguments{
\item{learners}{a list of learners}
}
\value{
A stacking model for binary outcomes.
}
\description{
Construct a stacked model
}
