% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_xgb.R
\name{binary_xgb}
\alias{binary_xgb}
\title{Gradient Boosted Tree Classifier}
\usage{
binary_xgb(
  eta = 0.05,
  max_depth = 0:6,
  subsample = 1,
  colsample_bytree = 1,
  nrounds = 10000L,
  workers = 1
)
}
\arguments{
\item{eta}{learning rate.}

\item{max_depth}{maximum depth of a tree.}

\item{subsample}{subsample ratio of observations when constructing each
tree. Values less than 1 will speed up computation, and may reduce
overfitting.}

\item{colsample_bytree}{subsample ratio of columns when constructing
each tree. Values less than 1 will speed up computation, and may reduce
overfitting.}
}
\value{
A gradient boosted tree model for binary outcomes,
with class 'learner'.
}
\description{
Construct a gradient boosted tree classifier using xgboost.
All parameters can be passed as single values, or as vectors. Performs
grid search over all combinations of provided values, using early stopping
to select the best number of boosting iterations.
}
