% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_learner_constructor}
\alias{is_learner_constructor}
\title{Check if Object is a Learner Constructor}
\usage{
is_learner_constructor(x)
}
\arguments{
\item{x}{object to check}
}
\description{
Check if Object is a Learner Constructor
}
