% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass_elasticnet.R
\name{multiclass_elasticnet}
\alias{multiclass_elasticnet}
\title{Multiclass elasticnet Regression}
\usage{
multiclass_elasticnet(
  alpha = seq(0, 1, length.out = 5),
  n_class = NA,
  workers = 1
)
}
\arguments{
\item{alpha}{sequence of alpha values to search over}

\item{workers}{integer number of cores to use for parallelized tuning}
}
\value{
A logistic elasticnet regression model for binary outcomes,
with class 'learner'.
}
\description{
Construct a elasticnet regression model with glmnet.
}
