% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_convex_combination.R
\name{regr_convex_combo}
\alias{regr_convex_combo}
\title{Convex Combination for Regression}
\usage{
regr_convex_combo()
}
\value{
A model that fits a weighted average of the inputs, with weights
all nonnegative and summing to 1.
}
\description{
Convex Combination for Regression
}
\details{
Does not use weights.
}
