% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_elasticnet.R
\name{regr_elasticnet}
\alias{regr_elasticnet}
\title{Elasticnet Regression}
\usage{
regr_elasticnet(alpha = seq(0, 1, length.out = 5))
}
\arguments{
\item{alpha}{sequence of alpha values to search over}
}
\value{
A logistic elasticnet regression model for continuous outcomes,
with class 'learner'.
}
\description{
Construct a elasticnet regression model with glmnet.
}
