% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_lgbm.R
\name{regr_lgbm}
\alias{regr_lgbm}
\title{Gradient Boosted Tree Regression}
\usage{
regr_lgbm(
  learning_rate = 0.05,
  num_leaves = 31,
  max_depth = -1,
  feature_fraction = 1,
  force_col_wise = FALSE,
  metric = "l2",
  nrounds = 10000L,
  workers = 1
)
}
\arguments{
\item{learning_rate}{learning rate.}

\item{max_depth}{maximum depth of a tree.}

\item{feature_fraction}{bagging_fraction ratio of columns when constructing
each tree. Values less than 1 will speed up computation, and may reduce
overfitting.}

\item{workers}{number of threads to use, to speed up training and prediction}
}
\value{
A gradient boosted tree model for continuous outcomes,
with class 'learner'.
}
\description{
Construct a gradient boosted tree regression using lightgbm
All parameters can be passed as single values, or as vectors. Performs
grid search over all combinations of provided values, using early stopping
to select the best number of boosting iterations.
}
