% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_sampmean.R
\name{regr_sampmean}
\alias{regr_sampmean}
\title{Sample Mean for Regression}
\usage{
regr_sampmean()
}
\value{
A model that predicts the weighted sample mean
}
\description{
Sample Mean for Regression
}
