% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_stacking.R
\name{regr_stacking}
\alias{regr_stacking}
\title{Stacking for Regression}
\usage{
regr_stacking(
  learners = list(elasticnet = regr_elasticnet(), gbt = regr_gbt())
)
}
\arguments{
\item{learners}{a list of learners}
}
\value{
A stacking model for continuous outcomes.
}
\description{
Construct a stacked model
}
