% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_stepforward.R
\name{regr_stepforward}
\alias{regr_stepforward}
\title{Forward Stepwise OLS Regression}
\usage{
regr_stepforward(workers = 1)
}
\value{
A linear regression model for continuous outcomes,
with class 'learner'.
}
\description{
Forward Stepwise OLS Regression
}
